chore(broker): Replace GSON with Jackson in BrokerCfg and GatewayCfg
Background:
The configuration objects are serialized to JSON and written to the log during startup.
Currently the behavior uses GSON. This is one of the few places where GSON is used.
In all other places we use Jackson. There is nothing special about using GSON here,
so it can and should be replaced with Jackson.

Scope
- Replace GSON with Jackson in BrokerCfg and GatewayCfg
- Factory method for JSON object writers with custom serializers for Duration (and other Java time objects)
  and DataSize objects

Out of Scope

Dependency to GSON in ExporterConfiguration - the class had a low test coverage and comparing
the default output of GSON vs. Jackson revealed that they are not identical, so changing the
serializer/deserializer here could alter behavior.